ADD AbstractPaginationFilter for NRPS/AGS use
Generalize out the NRPS PaginationFilter to AbstractPaginationFilter.
Turns out AGS uses the same pagination mechanism, so there should be a
lot of commonality between them.

This required a shared data structure for passing the link URLs to the
common functions. Each link can be described by an array composed of
'url' and a 'rel' keyword (if present). The data structure is just an
array of those links.

The parseLinkHeader() function takes the raw link header string and
spits out the array of links.

The createLinkHeader() function does the inverse, taking the array of
links and putting out a link header string.