UPDATE laravel 7 to 8
Interestingly, Laravel 7 is already no longer supported for bug fixes.
It looks like 7 and 8 are not considered LTS releases and won't have a
very long life. So we'll probably need to upgrade in another half a
year.

I wanted to use the bulk insert operations now available in laravel 8,
so it was a good tiem to upgrade anyways.

The largest breaking change is that database factories are now
individual classes. This required a large amount of small changes to the
factories and the test cases.