ADD ags PaginationFilter and LineitemsFilter
Rewrite the lineitem endpoint on the platform to the lineitem endpoint
on the shim.

The lineitems endpoint is at: /lti/ags/platform/<ags_id>

This should create a lot of specific lineitem endpoints at:
/lti/ags/platform/<ags_id>/lineitem/<ags_lineitem_id>

Like with NRPS, I would prefer the lineitems operations to be in bulk to
save on the amount of sql queries being done. This is now a lot easier
due to Laravel 8's bulk insertOrIgnore() function.

Rewrite lineitems pagination endpoints to lineitems endpoints on the
shim. Also similar to the NRPS pagination endpoint rewrite, we just
create new ags entries and link to them.

Honestly not sure if hiding the query parameters that the original
endpoint gives us is necessary, but erring on the side of hiding
information for now.