EDIT AccessToken verify now checks tool & scopes
A tool could access another tool's NRPS endpoint and obtain data that it
wasn't meant to have. So I've added a check to catch tools trying to
access another tool's endpoints.

Access token scope wasn't necessary before because we only had NRPS and
it only has one scope. If NRPS is advertised, we should be able to
safely assume that the NRP scope is supported. However, with the
introduction of AGS and its multiple scopes, we now need a way to
properly check the access token has the required scope to proceed with
an operation.

Since AGS has several scopes, some of which are interchangable depending
on the operation needed, as long as one of the required scopes is listed
in the access token, we consider it a pass. An example is the read only
and write allowed scopes. If we're only doing a read operation, then
either scope should allow access.

NrpsTest now checks that access tokens with incorrect tool or scope are
properly rejected.