EDIT ags now supports creating new lineitem
A POST request to the lineitems url should contain a single lineitem
entry. The lineitem entry should be created by the platform and returned
enhanced with the 'id' field containing the new lineitem's url.

Decided to group Platform lineitem operations into a single
PlatformAgsLineitems class, since it was growing large. We will still
have the single PlatformAgs class but it's called by the controller and
will direct the request to the appropriate class to handle it.

Also put placeholder for all the lineitem operations.

A basic test was created to test this but it's lacking a lot edge case
coverage. For example, it doesn't make sure that access tokens are
properly checked. Or checking for media type for that matter.

The most complicated part of this is ensuring that the AGS endpoint has
the correct scope for the operation. Was previously using the lineitem
readonly scope and hence my initial requests were rejected. The shim
should now be properly enforcing scope too, where if write permission
was not advertised in the launch, the request should be rejected before
we even proxy the request.

Tried to farm out scope resolution to the Ags model. Not really happy
with the code area scopes. The problem is that when verifying scopes,
there could be multiple equivalent scopes that is safe to use. But when
requesting scopes, it's best to only request 1 scope per access token
(some LMS not happy with multiple scopes requests).

Didn't really pay attention to media type before, but the shim should be
sending the correct media types for AGS.

phpunit.xml was modified so that tests will stop on the first failure.
Apparently, stop on failure was false before but had no effect because
the functionality was not implemented. Development is easier for me when
it stops on the first failure, so I've enabled stop on failure.

Added a render() method to LtiException that Laravel will call to handle
the exception response. Previously, I was manually catching LtiException
in controllers. With this change, I no longer have to catch LtiException
manually in new controllers. This is needed because by default,
exceptions result in a 500 error but LtiExceptions should be 400 errors.