EDIT ags now supports editing a single lineitem
A PUT request to a single lineitem's url should edit the lineitem.

I noticed that Canvas doesn't allow you to change the scoreMaximum. If
you try to PUT an updated scoreMaximum, it'll succeed and even return a
lineitem with the new scoreMaximum, but if you GET the lineitems again,
you'll find that old scoreMaximum is still there. I don't think this is
the fault of shim, so just going to leave it alone.

Most of the elements are already in place from creating a new lineitem,
so this didn't need much work.

Basic test for editing lineitem was added to AgsTest.