EDIT ags can now delete single lineitem
A DELETE request to a lineitem url should delete the lineitem.

The Canvas documentation states that the delete request's response
should be a lineitem, but it's actually an empty response as you'd
expect, making the implementation simpler.

I've also tweaked the response codes. Where creating a new lineitem will
return 201 CREATED and deleting a lineitem returns a 204 NO CONTENT.

The 204 is part of Laravel's empty response and I didn't want to change
it anyways.