EDIT AgsTest now tests lineitem filters/pagination
AgsTest now tests to make sure that lineitems filters as defined in the
spec are passed through to the original platform URL. It also tests that
pagination urls in the links header are converted to shim urls.

Now that I know the Result service will also have pagination, I've
renmaed the PaginationFilter to LineitemPaginationFilter since it only
deals with Lineitems.

Minor bits of refactoring that resulted from issues that propped up when
writing the tests.