EDIT ags result service proxying but no filter
AGS Result Service requests are now properly proxied. However, no
filtering has been implemented. Also, no filter queries are passed on.

There is now a PlatformResult and ToolResult to handle AGS result calls
as a platform/tool. PlatformAgs is still the only one that AgsController
calls which directs the call to the appropriate platform service.

Since there's only one call for both the platform and tool side, we can
simplify it a bit compared to the lineitem implementations.

To create the lineitem result URL, we need to attach "/results" to the
end of the URL path. This requires using the URL manipulation library
since we might have to add on queries and such. So the original
addParamsToUrl() method was extended to allow adding to path and renamed
addToUrl()

There is only one scope for results and it is read only.

Split AgsTest up. The previous tests are all AgsLineitemTest and the new
result service tests are in AgsResultTest. AgsResultTest required adding
a factory for AgsLineitem.