EDIT ags result filtering now works
Added a new table ags_results to track results. Each result has its own
url, so we had to do the same type of url translation as we've been
doing for NRPS and AGS.

There's only one filter, the ResultsFilter that takes care of all 3 of
the fields that needs to be filtered:

* id: This field is just the url to get info on a specific result.
* scoreOf: This is the url to the lineitem.
* userId: This is the 'sub' param from launch that identifies users.

For userId, if we encounter an unknown user, we error out. This probably
shouldn't happen in real use, since presumably, anyone who has a grade
would have passed through LTI beforehand. We technically could
create a real/fake user mapping with just the userId if this causes
issues.

While the individual result url is present, the actual call returns "Not
Implemented" error. Only enough to generate the url itself was
implemented. This can be done after finishing the filter param
passthrough.

Strange issue occurred when I tried to set a belongsTo() relationship in
the AgsResult model to AgsLineitem. If the function was named
ags_lineitem(), it did not work. If the function was named
ags_lineitems(), it worked. Not sure why the s is required when other
models, such as in LtiFakeUser, use singular forms for the relationship
definitions. Maybe the inverse hasMany() or hasOne() relationship
affects it.

AgsResultTest was modified to expect the filters. Also now properly
checks to see if token is enforced.