EDIT ags result filter queries passthrough
GET requests to the result url can contain two queries used for
filtering:

* user_id: limit results to a single user
* limit: limit number of results returned

limit can just be passed through as is. The resulting pagination link
header still needs to be dealt with though.

user_id needs to be rewritten. Since user_id comes from the tool, it is
using the fake user id. We need to map it to the real user id in order
for the platform to understand it.

If we can't find the user in our database, then we send back an empty
result. Maybe we should error out instead but trying an empty result for
now.

Interestingly, Canvas currently seems to only understand numeric Canvas
user id right now. The documentation states that the lti user id should
also be acceptable but I couldn't get it to work. Maybe this will be
patched in a later update.

Fixed ToolResult access token requesting unnecessary scope.

Fixed PlatformLineitem's request completed log message should be a
notice() level, not info.