EDIT ags result pagination now working
Rewrite the link header so that the pagination urls are pointing to the
shim. This is the same pagination mechanism as previously implemented
for lineitems and NRPS.

The actual result url on the shim is a bit weird. Technically, getting a
single result isn't in the spec. However, Canvas does implement it since
it seems like a rather obvious extension of the spec. So we should
implement it too. This is what AgsResult tracks.

Each pagination link however, also creates an AgsResult entry. So the
shim url ends up looking the same. This means that some AgsResult shim
url calls will end up with a single result and some ends up with an
array of results.

To make it uniform, I've decided to put even the single result returns
into an array of results. This is because there's only a single media
type for results defined in the spec, and that media type expects you to
return an array anyways.

Given that this is undefined behaviour in the spec, I don't think
there's a "right" way to do this anyways.

Figured out why AgsResult couldn't access AgsLineitem through the
ags_lineitem property but had to use ags_lineitems. Turns out I named
the foreign key ags_lineitems_id. Renaming the foreign key to
ags_lineitem_id made it all work as it's supposed to.