EDIT ags score service now working
The score service takes a post request on the lineitem url with a
'/scores' appended to the path. The post request should create a result
for that lineitem.

The score contains a userId that we need to de-anonymize for submitting
to the platform.

The return data, at least for Canvas, is a url to the result. This is
the same as the ResultsFilter's id. I'm not sure what other platforms
will return as I can't seem to find reference to the "resultUrl"
parameter returned by Canvas in the spec. However, it might be described
by a newer version of the spec. I've recently noticed that the linked
public specs on the LTI site isn't the latest version on the closed
github repos.

Added a simple test fot the score endpoint.