EDIT gitlab ci to be more similar to auto devops
Experimented with Auto Devops again, unfortunately, was unable to get
deployment working. This is due to the shim's chart requiring helm 3,
but helm 3 isn't supported in the current auto-deploy-image we're using.
We're on auto-deploy-image v1 but we need v2 for helm 3. Unfortunately,
v2 also requires kubernetes v1.16+ and we're still on v1.14.

CONFIGURATION_GITLAB_* params are not needed for Auto Devops.

Use the AUTO_DEVOPS_CHART_REPOSITORY_NAME param to specify the helm repo
name, was previously just using a hard coded name "gitlab".

Remove image parameters passed into helm, as we'd prefer to use the
chart provided values.