ADD gradebook message type allowed through
AGS companion spec that allows launching from a gradebook result into
the tool. Figure this was a small spec that we could implement as an
exercise to figure out how easy it is to modify the current Launch code.
Since the Deep Link spec would probably require us to modify the Launch
more extensively.

Gradebook Messages have the message type "LtiSubmissionReviewRequest",
this commit allows launches with that message type through.