ADD GradebookMesssageFilter filters for_user claim
The for_user claim tells the tool which student's grade we want to look
at. The user who performed the launch (presumably the instructor) is
still specified in the normal launch user info.

This is a relatively simple filter since we just need to look up the
real user to grab the fake user.

Note that we're filtering the for_user claim even if it's not a
Gradebook Message. If other specs also use the for_user claim, then we
probably should split out the for_user filtering into its own filter
class.

Also modified the test to make sure this filtering is taking place.