ADD AddToUrlTrait refactor of uri operations
We were using the Uri library to add queries to URLs a lot. But I was
just copy and pasting the same code over to different models. This
obviously is less than optimal. So I've extracted out the function into
a trait that can be used by all the models. Now there's only one copy of
the code in AddToUrlTrait.