ADD launch presentation claim's return url support
The core spec defined a return url as part of the launch presentation
claim. The return url can be used by the tool to return to the platform
with a set of messages passed as queries. These message queries are:

* lti_msg - message meant to be shown to the user
* lti_errormsg - error message meant to be shown to the user
* lti_log - message that should be logged
* lti_errorlog - error message that should be logged

To hide the original return url, we replace the url with an endpoint on
the shim. For tracking this, a new 'return_urls' table was created. When
they visit the shim endpoint, we send a redirect to the original URL.
This redirect should include the message queries.

Since return url support isn't really part of the Launch, I've created a
new Core/ directory for it as appropriate to separate it.

For basic security, when we created the shim return url, we also create
an access token as part of the shim return url. This isn't really much
security, but at least it prevents the basic attack of just incrementing
the return url id to get all of our return urls.

The launch presentation was modified to pass through the return url.
I've also decided to pass through the locale, cause that probably isn't
secret anyways.

Associated tests were included for return url.