ADD SubmissionReviewFilter to AGS lineitem
Gradebook Message adds a lineitem property named 'submissionReview' that
indicates the given the lineitem supports a Gradebook Message launch.

Each submissionReview must have a 'reviewableStatus' property. An error
will be return if that is not present.

submissionReview contains an optional param called 'url' that can
override the launch's 'target_link_uri'. This isn't supported for now
since it requires a lot of bookkeeping. Instead, I'm just stripping the
submissionReview from lineitems that have this override url. An error
should be logged in lti log.

Note that since this override url is given by the tool, we need to have
this filter on lineitem create POST requests too, before we pass the
lineitem to the actual platform.

While tests were added to check that get lineitems were properly
filtered. The post request lineitem filter is not currently being
checked. However, the filter is being applied from manual testing.