FIX npm audit, axios needs to be upgraded
npm audit found a vulnerability with our current axios 0.19, so we
needed to upgrade to 0.21. Upon upgrading however, it broke our front
end compilation again, go figure.

It keeps on insisting installing a separate version of webpack from the
one pulled in by Laravel Mix. I'm guessing somewhere along the way, we
pulled in a requirement for the new webpack 5. Unfortunately, the
version of Laravel Mix we're on (v5) only supports webpack 4. So I've
now also upgraded Laravel Mix to v6, which does support webpack 5. This
required a few additional changes as specified in the Laravel Mix 6
upgrade guide.

Hopefully, this is the end of the front end compilation issues.