ADD vue-router, REFACTOR separate out admin ui
Add vue-router for front end admin ui routing. Previously, I was
thinking of the admin UI as a very simple one page affair.
Unfortunately, feedback has shown that we require more separation
between components, which necessitates the usage of a router.

I also want to reorganize the js side so that it's more clear what
belongs to the admin ui and what belongs to the midway ui. To that end,
I'm going to put all admin js in its own admin/ directory. I've also
named the controllers/blade views/layouts as admin related.

This is also preparation for switching authentication to Laravel
Sanctum, which should fix the having to login twice bug and provide an
easy way to authenticate midway api calls.

It should be noted that a lot of what I need to do is technically
available as a separate Laravel package. Laravel 8 deprecated the old UI
scaffolding I was using previously and introduced two new options. One
of them in particular, Laravel Jetstream, is vue/tailwindcss based and
would be a great choice... if it weren't for some issues.

When I tried out Laravel Jetstream, it broke a lot of things, but those
were fixable.  What was really odd is that it increased the test run
time by orders of magnitude (from seconds to hours). I'm pretty sure
something about the migrations added by Jetstream dramatically slowed
down the testing sqlite db's refresh, as the tests seem to run fine when
using the actual postgres db. As the extremely slow tests make quick
development impossible (and I really don't want to mess with Jetstream
migrations), Jetstream cannot be used.