EDIT midway ui now uses Sanctum token auth
The Sanctum token is generated by MidwayLaunch and passed on as a prop
to the midway vue component.

We can't generate a token that isn't attached to a user, so I've created
a midway api user that should be autocreated as needed, for generating
midway api tokens.

The only API call we're concerned about for midway is the list of users
we can lookup identities for. This was previously also passed as a prop
to the vue component. It has now been converted to an api call handled
by the new MidwayApiController. This list of users is now being handled
by a vuex store named lookup.

Axios must be configured to use the token in the authorization header
before any api calls take place.

It looks like Sanctum's token auth does not enforce CSRF protection. I'm
a bit confused at this. I thought CSRF protection was always enforced.
Will need to check if this is a security concern.

TODO: Only allow the token to access users for the course context and
tool they were issued for.