EDIT dockerfile no longer use deploy-env
I want to stop using the .env file and start relying on environment
variables instead for configuration overrides. So removing its inclusion
in the Dockerfile and also edited docker-compose.yml to take its place.

I found that the delay in LoginView wasn't long enough when using
docker-compose, so I've increased it. This is now a long enough delay
that it's noticable, so I've added a spinner to indicate to users that
it is working.