FIX ci broken by outdated helm can't read v2 chart
Helm 2 cannot read v2 charts. Need to upgrade to helm 3.

helm init no longer exists as a command
helm fetch is renamed to pull

Also no longer requires tiller.

There is an issue where nginx-ingress passes the clusterIP with an empty
string. Helm 3 complains that it is trying to overwrite an immutable
property. I added the fixes but it didn't work for an upgrade install
(only works on a fresh install). So I just manually deleted the services
with kubectl, hopefully that'll work now.