EDIT .gitlab-ci.yml switch to Gitlab AutoDevOps
Removed the really outdated heavily customized autodevops gitlab-ci and
tried to use the new one, which comes in an image apparently.

We need to override auto-deploy-image version. The default
auto-deploy-image is v1.0.7 and we need v2+ for helm 3.

I really would prefer not having a custom .gitlab-ci.yml anymore, but
changing the auto-deploy-image for our whole gitlab instance might break
stuff for other people, so this is unavoidable. The good news is that
this should be the only customization we need.