REMOVE lti config own_tool_id/own_platform_id
These two configuration values tells us how to find the shim's own
platform and tool info. Unfortunately, these ids cannot be counted on to
be consistent with the database's id when we run tests. This is due to
postgres not resetting id auto increment when rolling back transactions.
So when tests undo the changes they made and reseed the database, we end
up with different ids.

This wasn't an issue when we were using sqlite as the test database.
Unfortunately, Laradock's version of sqlite is quite old and doesn't
support upsert (need 3.24.0 but stuck on 3.22.0). So we had to switch to
using postgres for testing.

Since we can no longer rely on ids, I've switched to search by iss for
the shim platform entry and searching by client_id for the shim tool
entry. The new lti.own_tool_client_id config tells us what client id to
look for and lti.iss already exists.