EDIT LtiRealUser will now update info from NRPS
The previous implementation only added new users that come in from NRPS.
If user information changes, we do not update the info. This is a
different behavior from the launch, which will update the real user
info. This is mainly because I wanted to stick to one query bulk updates
if possible for performance reasons. Now with the introduction of bulk
operation abstractions in Laravel, namely the upsert(), this can be
accomplished easily.

The good news is that we are actually able to remove code with this, as
we no longer have to limit ourselves to only new users.

Note that only name, email, and student number are updateable.

Also added tests to check that new users that come in via a second NRPS
call are added and that information changes from a previous NRPS call
are updated in the database. Since this requires that we send different
fake NRPS data in sequence, the http faking had to be reworked to allow
that. A setupHttpFake() call now needs to be done for every test case in
NrpsTest, providing an option to override the data sent.