EDIT switch test db from sqlite to postgres
Laradock's version of sqlite is at 3.22.0. Sqlite support for upsert
requires at least 3.24.0. I tried upgrading sqlite manually, but it was
complicated enough that I don't think it's a sensible path to go down.

This will require you to delete Laradock's postgres data
(e.g.: rm -rf ~/.laradock/data/postgres ). As the updated Laradock
addition to create a test db in postgres does not get executed if there
is existing data. You can also create the db yourself, the database name
should be "default_test". I just took the configured postgres db name in
.env and appended a "_test". The same postgres user used to access the
db should be able to access the test db too.

The test database connection is configured in config/database.php as
pgsql_testing. phpunit.xml is configured to use this this connection.

Postgres does not reset autoincrement ids on transaction rollback,
unlike sqlite. So when we run test cases, we cannot count on predictable
ids. Database seeders and test cases needed to be fixed to account for
this.