EDIT Midway UserList search&sorts by platform user
The Midway user look up for instructors now properly work when sorting &
searching on the platform side (real users). Previously, sort & search
would still be performed on tool users (fake users) even if the platform
was selected.

This mostly involved a lot of troubleshooting and experimenting with
Laravel's query builder to figure out how to make it work.
Interestingly, it looks like the entire API call requires 3 queries:
1. The query we wrote with search/sort, notably, this does NOT bring in
   real users by default.
2. Count the total number returned.
3. Pull in data from the real users table.

So, in order to be able to search/sort on real user data, we need to
JOIN the lti_real_users table in step 1. Note that we joined
'lti_real_users' as 'lti_real_user' to make it match the model relation
dropping the 's' from the table name. Without this, we would have to
rename individual columns, e.g.: 'lti_real_users.name AS
lti_real_user.name'. This was my original approach but it caused issues
for step 2, as the count query generated by Laravel does not do this
renaming, causing a 'lti_real_user.name' missing column error.

Still need an error handling mechanism on UserList. Currently, errors
are silent.