EDIT LtiSession cols null, new platform_client_id
Currently, an LtiSession entry is made at the last step of the LTI
launch, where we have all the information available and can fill in all
the columns. However, experimenting with Deep Link resulted in a flow
where we needed to create an LtiSession entry at the first step of the
LTI launch, without a lot of the information we need to fill in all the
columns of LtiSession. So these columns have to be made nullable.

The "doctrine/dbal" package is required for schema change() operations.

I've seen mention that foreign key enforcement needs to be disabled when
we set nullable() on foreign keys. However, I didn't have an issue
running the migration without that step. Maybe the newer version of
doctrine/dbal know enough to do that for us.

Added new foreign key "platform_client_id". This used to be passed as a
value in the EncryptedState JWT. The previous control flow created an
LtiSession entry later in the process, so storing the PlatformClient in
LtiSession didn't make sense. However, now that we're creating an
LtiSession from the start, we can move all such state keeping into the
LtiSession. This also decreases the size of the EncryptedState JWT we
have to pass around, since all it needs to hold is the LtiSession key.

When migration is run on an existing database, it'll fail because the
new platform_client_id column is not nullable. So as part of the
migration, we need to initialize values for the column on existing data.
This resulted in a multi step column creation, first we need to create
platform_client_id as a nullable regular column, set a value for the
column for existing rows, then properly configure it as a foreign key.