ADD DeepLink launch overhaul of OIDC login
The Deep Linking spec is just a modification of the standard LTI launch.
One key difference is that there is configuration required to be
performed midlaunch by the user on the tool side. This is similar to how
our current launch's Midway page is used. This makes it difficult for
our old launch flow since we complete the entire launch with the
platform before starting launch with the tool.

So in order to implement Deep Linking, we're going to implement the flow
suggested in #4, as it goes through the entire launch stage by stage
with both the platform and tool at the same time.

So the new flow goes:

1. Platform --(login)-----------> Shim --(login)-----------> Tool
2. Tool     --(auth req)--------> Shim --(auth req)--------> Platform
3. Platform --(auth resp)-------> Shim --(auth resp)-------> Tool
4. Tool     --(deep link resp)--> Shim --(deep link resp)--> Tool

This commit implements just step 1.

Step 4 is deep link launch only, a regular launch stops at step 3. In a
deep link launch, the end of step 3 is where the user is expected to
interact with the tool to generate the response in step 4.

The current plan is to implement this new flow just for deep link for
now, and once that is working, we can replace the old launch flow with
this implementation.

The old code organization splits between the shim's tool and platform
side. This doesn't make sense anymore now that we're looking more at
launch stages, where each stage has to handle both the shim's tool and
platform sides. So with the deep link launch, I'm naming the files after
the stages themselves.

1. LoginHandler - handles the OIDC login stage
2. AuthReqHandler - handles the authorization request stage
3. AuthRespHandler - handles the authorization response stage
4. DeepLinkRespHandler - handles the deep link response stage