ADD db seeder for test cases, new base test case
The first test cases manually called model factories to setup the
database. This resulted in a lot of duplication of code and a bit of a
maintenance issue. To avoid that, I started using the default database
seeder that was built to seed the dev environment. This seeder was
mainly built for manual testing and was missing data that were easily
created with a manual LTI launch. This missing data is usually necessary
for the test cases and again, a lot of duplicated setup code was used to
address this.

In an attempt to solve the issues outlined above, I've now added another
database seeder dedicated for running test cases. This is paired with a
new base test case that new test cases should extend. This new base test
case will retrieve commonly used models from the database in its
setUp(), which should help reduce code duplication. The new base test
case is also configured to reset the database after each test and to run
the new test database seeder.

I've also changed the seeders to try to use example.(com/org/edu) for
all URLs. As sometimes, a broken test case could result in a real
request being sent out onto the test network if the http mocking wasn't
configured correctly.