ADD LoginTest to test DeepLink launch's OIDC login
LoginTest is the first test using the new db seeder for tests and
extending the new LtiBasicTestCase.

In Laravel 8, they introduced a new way to call database seeders in
tests. You can set the $seed and $seeder vars in the test class and the
RefreshDatabase trait will call them for you. Unfortunately, there seems
to be a subtlety between the old way of calling database seeders and the
new way that makes them incompatible with each other. This is explained
more in the LtiBasicTestCase comments. Suffice it to say that using the
new $seed var results in seeded data NOT being rolled back, which in
turn causes data conflict as there is leftover data.

If the other tests are also using the $seed way, this conflict doesn't
seem to cause any issues. However, all other ways of seeding data, e.g.
by calling seed() explicitly, throws an error due to the database
conflict.

For now, we're sticking with calling seed() explicitly. However, it
would be worth it to switch to using $seed, as it appears to be more
performant.