ADD DeepLink AuthReqHandler for launch auth req
Similar to the LoginHandler, AuthReqHandler handles the authorization
request stage of the LTI launch. The authorization request is sent by
the target tool to the shim, the handler checks it over and creates its
own authorization request to the originating platform.

I've realized that the old implementation of the LTI launch was quite
overzealous in its launch filtering, calling them in every stage on the
tool side. Since we're directly crafting the messages in login and auth
req, there's really no need to filter those messages in this new
implementation. Hopefully, this results in a speedup. The filters will
need to be brought back into place for the authorization response stage
that's coming up next.

Note that the nonce we receive from the target tool needs to be stored
so we can repeat it back in the auth resp message to the target tool.
While the nonce we generate will need to be checked for validity in the
auth resp message from the originating platform.

While we ourselves didn't use the lti_message_hint param in login, we
need to repeat it back if the originating platform did use it.
Similarly, if the target tool sent us a state, we need to repeat it back
in the next auth resp stage.

This means that LtiSession should now also store a nonce and a state (if
given). I've decided not to remove values from LtiSession that have been
used, to save a database call.

One thing to note is that since we're storing the LtiSession in
login_hint, errors surrounding login_hint will NOT have the same log
stream ID as the login stage. This is because the log stream ID is saved
in LtiSession, so if we can't recover the LtiSession, we do not know
what stream ID to use, which means a new random ID was assigned to the
log stream.