ADD tests for DeepLink's AuthReqHandler
Modified the test database seeder to also create an LtiSession entry.
Since this is a fully populated LtiSession (albeit with no state
history), I think individual test classes will still need to do some
initialization to make it match expected conditions.

Fixed JwksTest failing randomly sometimes due to order being taken into
account in comparisons when it shouldn't be. Unfortunately, since this
seems to be a random failure, I can't be sure if this is fixed since
it's hard to duplicate.

Fix the redirect_uri we send should not be midway, I mistook it for
target_link_uri. It's where the auth resp is sent, so it should be the
auth route.

Added saving the redirect_uri to LtiSession. We should use this
persisted value as the location to send the auth response. Previously,
we were making sure that the redirect_uri matches exactly what was
configured in the tool entry. However, this might be too strict, as I
suspect some tools might work by adding queries or fragments to it. Or
even change the path altogether. So I've relaxed it to just making sure
they're the same site.

Not much to say about the new tests itself, but it nice that it's all
shaping up to be more comprehensive yet simpler to maintain than the old
launch tests.