ADD UriUtil for common operations using League\Uri
To more properly do url comparisons, we should parse them using the
League\Uri library cause it does normalizations and such.

One common operation that I think we'll do a lot is to make sure that
links we're given are from the same site that issued them. So I've put
this operation in the new UriUtil class.

The same site check is simple, it just makes sure the scheme (e.g. http
or https), host, and port of the urls given it all match.

I wasn't sure I was using the Uri calls correctly, so threw a quick unit
test together for a sanity check.