ADD AuthRespTest for Deep Link AuthRespHandler
Since the Deep Link flow handles both tool/platform sides at once, this
test case mostly just combines the old platform/tool specific
AuthRespTest into one. No improvement in test coverage, but it should at
least cover the same things. Hopefully the test itself is easier to
maintain but the checkSuccessfulResponse() function is still too long.

Added new pre-retrieved data in the LtiBasicTestCase.

The old test case seemed to have some minor issues in setup, e.g. where
it used the shim platform for the iss when it should've used the
originating platform. It still worked in the end, since that's still a
valid platform, but it probably wasn't as accurate a test as it
should've been.