ADD DeepLinkFilter for launch filtering
Similar to other LTI services, Deep Link will put a claim into id_token
that needs to be filtered. Most of the claim data can be passed through
as is, so this initial implementation does just that.

Deep link does not use the usual 'LtiResourceLinkRequest' message type,
but its own 'LtiDeepLinkingRequest'. If the message type is not a deep
link message, the deep link claim will be dropped by the filter.

Added tests for deep link claim filtering to AuthRespTest.

Two things missing from this implementation: The deep link return URL
needs to be replaced by a shim url. Also, the 'data' claim is meant as
an opaque state value from the platform for deep links, this will also
need to be preserved.