ADD deep_links table, model, use by DeepLinkFilter
DeepLinkFilter now creates DeepLink entries in the database. This is
only to save the originating platform's return_url and state (if given
one).

The state is the deep link setting's "data" claim. This is an opaque
state string that's given by the platform which we need to pass back in
the deep link response. I didn't want to name the column "data" since
that seems to be far too generic, so I named the column "state" to
hopefully make it less problematic.

We'll give the tool our own opaque state in 'data', which is the same
idea as the encrypted JWT used to pass LtiSession around, except this
time storing the DeepLink id. Speaking of LtiSession, we might need a
foreign key to LtiSession in the DeepLink table if we need to use data
in the id_token.

The DeepLink model also provides a shim return url. This means I had to
scaffold a handler for the shim deep link return. This ReturnHandler
only returns a 501 Not Implemented right now.

Modified the test to check that DeepLink entries are properly created.
Also added a check to make sure that deep link claims are just dropped
if the message type isn't deep link. I thought about failing the launch
if there's a deep link claim but the message type isn't correct, but
decided to just leave a warning in the log and silently drop the deep
link claim.